alert_manager/requirements.txt
kafka-python
redis
psycopg2-binary
loguru
